Create functions.php
consolidated all of the functions from handle-et-data.php into one file. This still needs some cleanup. For example the updateListAttributes function needs to take an object and pull from that instead of lining out each attribute. I'll get around to it eventually. Also, the functions now reactivate the subscriber if they had previously unsubscribed. This was a serious flaw in the old function as subscribers would be added to the data extension, but would still be inactive.